modified Watchlist, to use inheritance fixing breaking changes
