+++ src/main/java/com/cbfacademy/apiassessment/Glossary.txt
@@ -1,3 +1,6 @@
- // @ResponseBody tells java the controller the object returned should be serialized to json
-    // @RequestBody tells java maps to HttpRequestBody automatically allowing deserialization
-    
+@Component indicates a class is a spring Component and is detected and registered as a spring bean when component scanning takes place
+
+@PathVariable like a template literal in js, it takes the variable it is attached to and sends it to its representation in the curly braces i.e when GetMapping ("/{id}/{name}/") public methodName(@PathVariable Long id, String name)
+
+@ResponseBody tells java the controller the object returned should be serialized to json
+@RequestBody tells java maps to HttpRequestBody automatically allowing deserialization
