+++ src/main/java/com/cbfacademy/apiassessment/service/UpdateWatchlist.java
@@ -1,0 +1,37 @@
+// package com.cbfacademy.apiassessment.service;
+
+// import java.util.UUID;
+
+// import com.cbfacademy.apiassessment.model.CreateWatchlist;
+// import com.cbfacademy.apiassessment.model.Watchlist;
+
+// // public Student update(Long id, Student student) {
+// //     student.setId(id);
+// //     Student oldStudent = repository.replace(id, student);
+// //     return oldStudent == null ? null : student;
+// public class UpdateWatchlist {
+
+//     private WatchlistService service;
+//     private CreateWatchlist CreateWatchlist;
+
+//     public Watchlist update(CurrentWatchlist currentWatchlist){
+//         //  Watchlist updatedWatchlist = new Watchlist();
+//         CurrentWatchlist updatedWatchlist = new currentWatchlist();
+//         // create a variable called current watchlist. its values populated with items returned back from watchlist;
+//         updatedWatchlist.setSymbol(currentWatchlist.getSymbol());
+//         updatedWatchlist.setOwned(currentWatchlist.isOwned());
+//         updatedWatchlist.setStatus(currentWatchlist.getStatus());
+//         updatedWatchlist.setCurrency(currentWatchlist.getCurrency());
+//         updatedWatchlist.setDatePurchased(currentWatchlist.getDatePurchased());
+//         updatedWatchlist.setUnitsOwned(currentWatchlist.getUnitsOwned());
+//         updatedWatchlist.setProfit(currentWatchlist.getProfit());
+//         updatedWatchlist.setOpen(currentWatchlist.getOpen());
+//         updatedWatchlist.setClose(currentWatchlist.getClose());
+//         updatedWatchlist.setIntradayHigh(currentWatchlist.getIntradayHigh());
+
+//         // UpdatedWatchlist response = new UpdateList(currentWatchlist, updatedWatchlist);
+
+//     }
+
+    
+// }
