+++ src/main/java/com/cbfacademy/apiassessment/service/UserWatchlist.java
@@ -1,11 +1,14 @@
 package com.cbfacademy.apiassessment.service;
 
+import org.springframework.stereotype.Component;
+
+import com.cbfacademy.apiassessment.model.CreateWatchlist;
 import com.cbfacademy.apiassessment.model.Watchlist;
 
-
+@Component
 public class UserWatchlist {
-     public Watchlist create(Watchlist copyOfWatchlist){
-        Watchlist copy = new Watchlist();
+     public CreateWatchlist create(CreateWatchlist copyOfWatchlist){
+        CreateWatchlist copy = new CreateWatchlist(null, null);
         copy.setSymbol(copyOfWatchlist.getSymbol());
         copy.setOwned(copyOfWatchlist.isOwned());
         copy.setStatus(copyOfWatchlist.getStatus());
