+++ src/main/java/com/cbfacademy/apiassessment/userInteractions/UserInput.java
@@ -3,6 +3,7 @@ package com.cbfacademy.apiassessment.userInteractions;
 import java.time.LocalDate;
 
 import com.cbfacademy.apiassessment.exceptions.InvalidInputException;
+import com.cbfacademy.apiassessment.model.CreateWatchlist;
 import com.cbfacademy.apiassessment.model.Watchlist;
 
 // uses the details that came from userInteractions class and converts them into usable watchlist object that can be used when in Post and put requests.
@@ -13,27 +14,27 @@ public class UserInput {
         this.userInteractions = userInteractions;
     }
     
-    protected Watchlist useDetails() throws InvalidInputException{
-        Watchlist watchList = new Watchlist();
+    protected CreateWatchlist useDetails() throws InvalidInputException{
+        CreateWatchlist createNewList = new CreateWatchlist(null, null);
         
 
         // watchList.setStockName(userInteractions.inputStockName()); we removed the setter as the property will be initialized on creation an will not be able to be edited after
-        watchList.setSymbol(userInteractions.inputSymbol());
-        watchList.setOwned(userInteractions.userOwnership());
-        watchList.setStatus(userInteractions.inputStatus());
-        watchList.setCurrency(userInteractions.userCurrency());
+        createNewList.setSymbol(userInteractions.inputSymbol());
+        createNewList.setOwned(userInteractions.userOwnership());
+        createNewList.setStatus(userInteractions.inputStatus());
+        createNewList.setCurrency(userInteractions.userCurrency());
         LocalDate date = userInteractions.userPurchaseDate();
-        watchList.setDatePurchased(date);
-        watchList.setUnitsOwned(userInteractions.userUnits());
-        watchList.setProfit(userInteractions.userProfit());
-        double pointsChange = watchList.getClose() - watchList.getOpen();
-        watchList.setPointsChange(pointsChange);
-        watchList.setOpen(userInteractions.openingPrice());
-        watchList.setClose(userInteractions.closingPrice());
-        watchList.setIntradayHigh(userInteractions.dailyHigh());
+        createNewList.setDatePurchased(date);
+        createNewList.setUnitsOwned(userInteractions.userUnits());
+        createNewList.setProfit(userInteractions.userProfit());
+        double pointsChange = createNewList.getClose() - createNewList.getOpen();
+        createNewList.setPointsChange(pointsChange);
+        createNewList.setOpen(userInteractions.openingPrice());
+        createNewList.setClose(userInteractions.closingPrice());
+        createNewList.setIntradayHigh(userInteractions.dailyHigh());
         userInteractions.closeScanner();
 
-        return watchList;
+        return createNewList;
     }
 
 }
